moves render_in_frontend function to decorators module, removes unused update_value_if_change function
